binanceBotTestnet/requirements.txt
APScheduler==3.10
pytz
python-binance==1.0.28
streamlit>=1.18.0
pandas
ccxt
python-telegram-bot
streamlit-authenticator
